Refactor - introducing MVC packaging
