+++ src/main/java/com/productmanager/entity/Drink.java
@@ -1,4 +1,4 @@
-package com.productmanager.data;
+package com.productmanager.entity;
 
 import java.math.BigDecimal;
 import java.time.LocalTime;
