+++ src/main/java/com/productmanager/entity/ProdManException.java
@@ -1,4 +1,4 @@
-package com.productmanager.data;
+package com.productmanager.entity;
 
 public class ProdManException extends Exception{
     public ProdManException() {
